+++ Assets/Scripts/Unity/Behaviours/GameManager.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 using Ventura.GameLogic;
 using Ventura.Util;
 
-namespace Ventura.Behaviours
+namespace Ventura.Unity.Behaviours
 {
     public class GameManager : MonoBehaviour
     {