+++ Assets/Scripts/Unity/Behaviours/MapManager.cs
@@ -1,11 +1,11 @@
 using TMPro;
 using UnityEngine;
 using Ventura.GameLogic;
-using Ventura.Graphics;
+using Ventura.Unity.Graphics;
 using Ventura.Util;
 
 
-namespace Ventura.Behaviours
+namespace Ventura.Unity.Behaviours
 {
 
     public class MapManager : MonoBehaviour